working on lexer/parser implementation
