+++ Gruntfile.js
@@ -16,6 +16,8 @@ module.exports = function(grunt) {
 				src: [
 					'src/wrapper/head.js.part', 
 					'src/error_handling/jSQL_Error.js',
+					'src/error_handling/jSQL_Lexer_Error.js',
+					'src/error_handling/jSQL_Parse_Error.js',
 					'src/error_handling/error_handling.js',
 					'src/data_types/jSQLDataTypeList.js',
 					'src/table/jSQLTable.js',
@@ -26,7 +28,10 @@ module.exports = function(grunt) {
 					'src/query_types/jSQLSelectQuery.js',
 					'src/query_types/jSQLUpdateQuery.js',
 					'src/query_types/jSQLCreateQuery.js',
+					'src/lexer/jSQLLexer.js',
 					'src/parser/jSQLParseQuery.js',
+					'src/parser/jSQLParseCreateTokens.js',
+					'src/parser/jSQLParseInsertTokens.js',
 					'src/parser/jSQLWhereClause.js',
 					'src/persistence/API.js',
 					'src/persistence/persistenceManager.js',
