+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jsql-official",
-  "version": "2.9.2",
+  "version": "3.0.0",
   "description": "A persistent SQL database.",
   "main": "jSQL.min.js",
   "directories": {
